refactor(Demo): improve the folder structure
