+++ demo/src/formly/wrappers/panel.wrapper.ts
@@ -1,5 +1,5 @@
 import { Component, ViewContainerRef, ViewChild } from '@angular/core';
-import { FieldWrapper } from '../src/core/templates/field.wrapper';
+import { FieldWrapper } from 'ng-formly';
 @Component({
   selector: 'formly-wrapper-panel',
   template: `
