+++ demo/src/services/index.ts
@@ -1,0 +1,2 @@
+export { SuperHerosService } from './super-heros';
+export { ValidationService } from './validation.service'
